src/celery_app/requirements.txt
celery==5.3.6
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
flower==2.0.1
